Correctness ratchet: strict clippy + panic hygiene + dep checks (#106)
Refs #105.

+++ backend/llm/src/cache_forensics.rs
@@ -16,6 +16,7 @@ use serde_json::json;
 static FORENSIC_DIR: OnceLock<PathBuf> = OnceLock::new();
 static CALL_COUNTER: AtomicU64 = AtomicU64::new(0);
 
+#[derive(Clone, Copy)]
 pub struct RequestLog<'a> {
     pub call_id: u64,
     pub character: Option<&'a str>,
@@ -30,15 +31,16 @@ pub struct RequestLog<'a> {
     pub rid: Option<&'a str>,
 }
 
+#[derive(Clone, Copy)]
 pub struct ResponseLog<'a> {
     pub call_id: u64,
     pub model: &'a str,
     pub character: &'a str,
     pub call_type: &'a str,
-    pub input_tokens: u32,
-    pub output_tokens: u32,
-    pub cache_read_tokens: u32,
-    pub cache_creation_tokens: u32,
+    pub input_tokens: u64,
+    pub output_tokens: u64,
+    pub cache_read_tokens: u64,
+    pub cache_creation_tokens: u64,
 }
 
 /// Enable cache forensics.  Call once at startup with the cache directory.
@@ -133,8 +135,8 @@ pub fn notify_anomaly(
     character: &str,
     anomaly: &str,
     call_type: &str,
-    cache_read: u32,
-    cache_write: u32,
+    cache_read: u64,
+    cache_write: u64,
 ) {
     if !is_enabled() {
         return;
